feat(predict-cart): add predict trackCartWithItems: SUITEDEV-14056

+++ Predict/EMSCartItemProtocol.h
@@ -7,8 +7,8 @@
 
 - (NSString *)itemId;
 
-- (int)price;
+- (double)price;
 
-- (int)quantity;
+- (double)quantity;
 
 @end
